+++ src/app/views/discover/communities/communities.route.ts
@@ -3,5 +3,5 @@ import { RouteRecordRaw } from 'vue-router';
 export const routeDiscoverCommunities: RouteRecordRaw = {
 	name: 'discover.communities',
 	path: '/communities',
-	component: () => import('./communities.vue'),
+	component: () => import('./RouteDiscoverCommunities.vue'),
 };
